@@ -3,7 +3,7 @@ package com.ead.course.dtos;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 @Data
 public class LessonDto {
