@@ -2,7 +2,7 @@ package com.ead.course.dtos;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
 
 @Data
